apps/backend/pyproject.toml
[project]
name = "voice-flow-backend"
version = "0.1.0"
description = "Voice Flow backend API."
readme = "../../README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.0.10",
    "livekit-api>=1.0.5",
    "phonenumbers>=8.13.0",
    "psycopg[binary]>=3.2.9",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
